Use correct type (int64_t) for ApexFile's version
.. and trust versions stored in MountedApexDatabase.

Bug: 381174544
Test: ApexTestCases

+++ apexd/apex_database.h
@@ -36,16 +36,16 @@ class MountedApexDatabase {
   // Stores associated low-level data for a mounted APEX. To conserve memory,
   // the APEX file isn't stored, but must be opened to retrieve specific data.
   struct MountedApexData {
-    int version = 0;        // APEX version for this mount
-    std::string loop_name;  // Loop device used (fs path).
-    std::string full_path;  // Full path to the apex file.
+    int64_t version = 0;      // APEX version for this mount
+    std::string loop_name;    // Loop device used (fs path).
+    std::string full_path;    // Full path to the apex file.
     std::string mount_point;  // Path this apex is mounted on.
     std::string device_name;  // Name of the dm verity device.
     // Whenever apex file specified in full_path was deleted.
     bool deleted = false;
 
     MountedApexData() = default;
-    MountedApexData(int version, const std::string& loop_name,
+    MountedApexData(int64_t version, const std::string& loop_name,
                     const std::string& full_path,
                     const std::string& mount_point,
                     const std::string& device_name)
